Fix Installation Bug and Restructure Server Hooks
